Update Globals and Rasters

There is also some part of the simulation state that can not be changed inside a transition functions and should be therefore modified between two refapply! calls, namly the globals and rasters.

Globals

You can add a new or additional value to the globals struct with:

Vahana.set_global! — Function

set_global!(sim::Simulation, name, value)

Set the value of the field name of the globals struct for simulation sim.

set_global! must not be called within a transition function.

See also create_simulation, mapreduce, push_global! and get_global

Vahana.push_global! — Function

push_global!(sim::Simulation, name, value)

In the case that a field of the globals struct from the Simulation constructor is a vector (e.g. for time series data), push_global! can be used to add a value to this vector, instead of writing set_global!(sim, name, push!(get_global(sim, name), value).

push_global! must not be called within a transition function.

See also create_model, mapreduce, set_global! and get_global

Base.mapreduce — Function

mapreduce(sim, f, op, ::Type{T}; [kwargs ...])

Calculate an aggregated value, based on the state of all agents or edges of type T.

f is applied to all of these agents or edges and then op is used to reduce (aggregate) the values returned by f.

mapreduce is calling Base.mapreduce, f, op and kwargs are passed directly to mapreduce, while sim and T are used to determine the iterator.

Warning

The state of the globals struct (the globals argument of the create_simulation function, and of rasters can not be changed inside of a transition function, as a transition function is calculated on a per agent basis with many evaluations in parallel, and changes to the global layer must be a single, synchronized operation.

Raster

Even if it looks like rasters are just nodes in the graph, and the type of those nodes must be registered like the types of agents via register_agenttype!, they have a special property: In a parallel simulation the state of the raster nodes are synchronized, so that the complete state of a raster is available on all ranks.

So to update the values of a raster, you can use one of the following functions:

Vahana.calc_raster — Function

calc_raster(sim, raster::Symbol, f, f_returns::DataType, accessible::Vector{DataType})

Calculate values for the raster raster by applying f to each cell ID of the cells constructed by the add_raster! function.

f_returns must be the type returned by the function f. There must be an implementation of the zero function for this type, and zero(returntype) | f(id) must be equal to f(id).

accessible is a vector of Agent and/or Edge types. This vector must list all types that are accessed directly (e.g. via agentstate or indirectly (e.g. via neighborstates in the transition function.

If the results of calc_raster depend only on the state of the cells (as in the following example) and all cells have the same type, calc_rasterstate can be used as concise alternatives.

Returns a n-dimensional array (with the same dimensions as raster) with those values.

Example:

The following code from a "Game of Life" implementation generates a boolean matrix indicating which cells are alive (and therefore maps the internal graph structure to the usual representation of a cellular automaton):

    calc_raster(sim, :raster, id -> agentstate(sim, id, Cell).active, Bool, [ Cell ]) 

Can be only called after finish_init!.

See also add_raster! and calc_rasterstate

Vahana.calc_rasterstate — Function

calc_rasterstate(sim, raster::Symbol, f, f_returns::DataType, ::Type{T})

Combined calc_raster with agentstate for the cells of the raster.

Calculate values for the raster raster by applying f to the state of each cell.

f_returns must be the type returned by the function f. There must be an implementation of the zero function for this type, and zero(returntype) | f(state) must be equal to f(state).

Returns a n-dimensional array (with the same dimensions as raster) with those values.

Example:

Instead of

    calc_raster(sim, :raster, id -> agentstate(sim, id, Cell).active, Bool, [ Cell ]) 

it also possible to just write

    calc_rasterstate(sim, :raster, c -> c.active, Bool, Cell)

Can be only called after finish_init!.

See also add_raster! and calc_rasterstate