Plots

Visualization is not the main focus of Vahana, but two rudimentary auxiliary functions exist for creating plots with Makie.

Vahana.plot_globals — Function

plot_globals(sim, names::Vector{Symbol})

Creates a Makie lineplot with one line for each global in names, wherby those global values must be Vectors (or Iterable).

Returns figure, axis, plots. plots is a vector of Plots, which one plot for element of names.

Info

plot_globals is only available when a Makie backend is imported by the client.

See also push_global!, get_global

Missing docstring.

Missing docstring for create_graphplot. Check Documenter's build log for details.