Transition Function

As mentioned in the create_model documentation, the model struct is created dynamically incl. a dynamically name for this struct. E.g. in the "Game of Life" Example this type is Vahana.var"Game of Life"{Params, Globals}. All of this types are subtypes of the abstract type Simulation.

Vahana.apply! — Function

apply!(sim, func, call, read, write; add_existing)

Apply the transition function func to the simulation state.

call must be a collection of agent types, read and write must be collections of agent and/or edge state types.

call determines for which agent types the transition function func is called. Within the transition function, an agent has access to the state of agents (including its own state) and to edges only if their types are in the read collection. Accordingly, the agent can change its own state and/or create new agents or edges only if their types are in the write collection.

Assume that T is an agent type that is in call. In case T is also in read, the transition function must have the following signature: transition_function(agent::T, id, sim:), where the type declaration of agent is optional if call contains only a single type. If T is not in read, it must have the signature transition_function(::Val{T}, id::AgentID, sim::Simulation).

If T is in write, the transition function must return either an agent of type T or nothing. If nothing is returned, the agent will be removed from the simulation, otherwise the agent with id id will get the returned state after the transition function was called for all agents.

When an agent or edge state type is in write, the current agents and edges of that type are removed from the simulation. Usually the agents are read by returning them from the transition function. If you want keep the existing agents/edges for a specific type, you can add this type to the optional add_existing collection.

See also apply

Vahana.apply — Function

apply!(sim, func, call, read, write; add_existing)

Call apply! with a copy of the simulation so that the state of sim itself is not changed.

Can be very useful during development, especially if Vahana is used in the REPL. However, for performance reasons, this function should not be used in the final code.

Returns the copy of the simulation.

See also apply!

Tip

The agent types must be immutable, but in most cases the agent returned by a transition function will have a different state than the agent specified as a parameter. If only one field is changed, we still need to copy all other fields. The Setfield.jl package can be very useful in this case.

Inside a transition function the following functions can be used to access the state of the simulation:

Globals and Parameters

Vahana.param — Function

param(sim::Simulation, name)

Returns the value of the field name of the params struct from the Simulation constructor.

See also create_model

Vahana.get_global — Function

get_global(sim::Simulation, name)

Returns the value of the field name of the globals struct for simulation sim.

See also create_simulation, set_global! and push_global!

Get Agent(state)

Vahana.agentstate — Function

agentstate(sim, id::AgentID, ::Type{T})

Returns the state of an agent of type T.

In the case where the type T is not determinable when writing the code (e.g. since there may be edges between agents of different types, the function edges may also return agentID of different agent types), agentstate_flexible must be used instead.

Warning

if agentstate is called with a Type{T} that does not match the type of the agent with id and the vahana assertions are disabled via enable_asserts, then it is possible that the state of an incorrect agent will be returned. When the assertions are active, there is a runtime check that the agent with the ID id has indeed the type T.

Vahana.agentstate_flexible — Function

agentstate_flexible(sim, id::AgentID)

Returns the state of an agent with the id, where the type of the agent is determined at runtime. If the type is known at compile time, using agentstate is preferable as this improves performance.

To get the states of all agents that are connected to an agent (with the agent as target) via edges of a specific edge type, the neighborstates functions can be useful (they are combing neighborids and agentstate).

Vahana.neighborstates — Function

neighborstates(sim::Simulation, id::AgentID, ::Type{E}, ::Type{A})

Returns the state of the agent with type A on the source of the edge of type E with agent id as target if E has the hint :SingleEdge, or a vector of these agent states otherwise.

If there is no edge with agent id as target, neighborstates returns nothing.

When the agents on the source side of the edges can have different types, and it is impossible to determine the Type{A} you can use neighborstates_flexible instead.

neighborstates is not defined if T has the hint :IgnoreFrom

See also apply!, checked, edges, neighborids, num_edges, has_edge and edgestates

Vahana.neighborstates_flexible — Function

neighborstates_flexible(sim::Simulation, id::AgentID, ::Type{E})

Returns the state of the agent on the source of the edge of type E with agent id as target if E has the hint :SingleEdge, or a vector of these agent states otherwise.

If there is no edge with agent id as target, neighborstates_flexible returns nothing.

neighborstates_flexible is the type instable version of neighborstates and should be only used in the case that the type of agent can not be determined.

neighborstates_flexible is not defined if T has the hint :IgnoreFrom.

See also apply!, checked, edges, neighborids, num_edges, has_edge and edgestates

Get Edge(state)

Vahana.edges — Function

edges(sim, id::AgentID, ::Type{E})

Returns the edge of type E with agent id as target if E has the hint :SingleEdge, or a vector of these edges otherwise.

If there is no edge with agent id as target, edges returns nothing.

edges is not defined if E has the hint :IgnoreFrom or :Stateless.

See also apply!, checked, neighborids, edgestates, num_edges, has_edge and neighborstates

Vahana.edgestates — Function

edgestates(sim, id::AgentID, ::Type{E})

Returns the state of the edge of type E with agent id as target if E has the hint :SingleEdge, or a vector of these states otherwise.

If there is no edge with agent id as target, edgestates returns nothing.

edgestates is not defined if E has the hint :Stateless.

See also apply!, checked, edges, neighborids, num_edges, has_edge and neighborstates

Vahana.neighborids — Function

neighborids(sim, id::AgentID, ::Type{E})

Returns the ID of the agent on the source of the edge of type E with agent id as target if E has the hint :SingleEdge, or otherwise a vector of the IDs of the agents on the source side of those edges.

If there is no edge with agent id as target, neighborids returns nothing.

neighborids is not defined if E has the hint :IgnoreFrom.

See also apply!, checked, edges, edgestates, num_edges, has_edge and neighborstates

Vahana.num_edges — Method

num_edges(sim, id::AgentID, ::Type{E})

Returns the number of edges of type E with agent id as target.

num_edges is not defined if T has the hint :SingleEdge

See also apply!, edges, neighborids, edgestates, has_edge and neighborstates

Vahana.has_edge — Function

has_edge(sim, id::AgentID, ::Type{E})

Returns true if there is at least one edge of type E with agent id as target.

has_edge is not defined if T has the :SingleEdge and :SingleType hints, with the exception that it has also the :IgnoreFrom and :Stateless hints.

See also apply!, edges, neighborids, edgestates, num_edges and neighborstates

For all the function like edges that returns nothing in the case that there is no edge with the agent as target, it can be useful to increase the readability of the code by using the checked function to test for nothing.

Vahana.checked — Function

checked(f, g, itr; kwargs...)

Calls g(f, itr; kwargs...), but only if itr != nothing.

As all the Vahana functions that access the edges of a specific agent can return nothing in the case, that there exist no incoming edge for this agent, it's often necessery to check this case.

Example:

Instead of writing

nids = neighborids(sim, id, Contact)
if nids != nothing
    foreach(nids) do nid
      add_edge!(sim, id, nid, Inform()
    end
end

you can use the checked function to write

checked(foreach, neighborids(sim, id, Contact)) do nid
    add_edge!(sim, id, nid, Inform())
end

Add Agents/Edges

The following functions from the [Initialization] section (initialization.md) that add new agents or edges to a simulation can also be used inside a transition function, namely:

add_agent! and add_agents!

add_edge! and add_edges!

connect_raster_neighbors!

move_to!